Added compiler flag for better warnings. Changed FS::EXT2 attribute types from u32 to i32.

+++ app-c/src/main/app.cc
@@ -3,7 +3,7 @@
 
 i32 main(i32 argc, i8 *argv[]) {
 
-    for (i8 i = 0; i < argc; ++i) {
+    for (i32 i = 0; i < argc; ++i) {
         printf("%d %s\n", i, argv[i]);
     }
